+++ src/models/UsedClassesFromJS.ts
@@ -1,0 +1,17 @@
+export interface IUsedClasses {
+  imports: IImports[];
+  classes: IClasses[];
+}
+
+export interface IImports {
+  stylesFilename: string;
+  importVariable: string;
+  sourceFile: string;
+}
+
+export interface IClasses {
+  usedClassName: string;
+  importStyleName: string;
+  stylesFilename: string;
+  sourceFile: string;
+}