+++ src/models/index.ts
@@ -1,0 +1,2 @@
+export type { IUsedClasses, IImports, IClasses } from './UsedClassesFromJS';
+export type { ISelectors, IPosition } from './DefinedClassesFromSCSS';